Move string truncation helpers into codex-utils-string
