+++ codex-rs/core/src/lib.rs
@@ -78,7 +78,6 @@ mod stream_events_utils;
 pub mod test_support;
 mod text_encoding;
 pub use codex_login::token_data;
-mod truncate;
 mod unified_exec;
 pub mod windows_sandbox;
 pub use client::X_RESPONSESAPI_INCLUDE_TIMING_METRICS_HEADER;
