+++ codex-rs/core/src/rollout/recorder_tests.rs
@@ -10,8 +10,8 @@ use codex_protocol::protocol::SandboxPolicy;
 use codex_protocol::protocol::TurnContextItem;
 use codex_protocol::protocol::UserMessageEvent;
 use pretty_assertions::assert_eq;
+use std::fs;
 use std::fs::File;
-use std::fs::{self};
 use std::io::Write;
 use std::path::Path;
 use std::path::PathBuf;
