+++ codex-rs/protocol/Cargo.toml
@@ -16,6 +16,7 @@ codex-execpolicy = { workspace = true }
 codex-git-utils = { workspace = true }
 codex-utils-absolute-path = { workspace = true }
 codex-utils-image = { workspace = true }
+codex-utils-string = { workspace = true }
 icu_decimal = { workspace = true }
 icu_locale_core = { workspace = true }
 icu_provider = { workspace = true, features = ["sync"] }
