+++ codex-rs/utils/output-truncation/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "output-truncation",
+    crate_name = "codex_utils_output_truncation",
+)
